Further Setup
Created showall.php that goes to show all slide when clicked. also link in bottom bit has link for show all so that show all can show results.

Adjusted some css, but didnt do much because this is a practice and I am not worried about aethetics

Also fixed bottom bit because it was missing a  php bracket ending.